moved enemy scripts into folders
