+++ .idea/deploymentTargetDropDown.xml
@@ -7,11 +7,11 @@
         <deviceKey>
           <Key>
             <type value="VIRTUAL_DEVICE_PATH" />
-            <value value="C:\Users\armin\.android\avd\Pixel_2_API_30.avd" />
+            <value value="C:\Users\grace\.android\avd\Pixel_2_API_31.avd" />
           </Key>
         </deviceKey>
       </Target>
     </targetSelectedWithDropDown>
-    <timeTargetWasSelectedWithDropDown value="2023-11-12T00:31:32.233232900Z" />
+    <timeTargetWasSelectedWithDropDown value="2023-11-12T23:36:23.829208900Z" />
   </component>
 </project>
