+++ local.properties
@@ -4,5 +4,5 @@
 # Location of the SDK. This is only used by Gradle.
 # For customization when using a Version Control System, please read the
 # header note.
-#Sun Nov 12 17:19:53 EST 2023
-sdk.dir=/Users/RayannLiang/Library/Android/sdk
+#Sun Nov 12 18:36:17 EST 2023
+sdk.dir=C\:\\Users\\grace\\AppData\\Local\\Android\\Sdk
